+++ app/Http/Controllers/api/Student/StudentListController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\api;
+namespace App\Http\Controllers\api\Student;
 
 use App\Http\Controllers\Controller;
 use App\Service\StudentListService;
