+++ tests/Feature/Student/StudentListControllerTest.php
@@ -6,7 +6,7 @@ use Tests\TestCase;
 use App\Models\Resume;
 use App\Models\Student;
 use App\Models\Tag;
-use App\Http\Controllers\api\StudentListController;
+use App\Http\Controllers\api\Student\StudentListController;
 use App\Service\StudentListService;
 use Tests\Fixtures\Students;
 use Tests\Fixtures\Resumes;
